10 bug fixes, 31 regression tests, type error cleanup,
migration consolidation, and old test suite removal.

+++ lib/services/image-cache.ts
@@ -148,7 +148,7 @@ export class ImageCache {
 						format: 'JPEG',
 						quality: 0.85,
 					});
-					buffer = Buffer.from(converted);
+					buffer = Buffer.from(converted as Uint8Array);
 				} catch (_heicError) {
 					structuredConsole.warn(`${LOG_PREFIX} HEIC conversion failed, using original`);
 				}
